Short and Long field should also take Short and Long objects as their defaults... (Typo fix)
git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221577 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/turbine/services/intake/model/LongField.java
@@ -85,7 +85,7 @@ public class LongField
     }
 
     /**
-     * Sets the default value for an Float Field
+     * Sets the default value for an Long Field
      *
      * @param prop Parameter for the default values
      */
@@ -98,7 +98,7 @@ public class LongField
             return;
         }
 
-        defaultValue = new Float(prop);
+        defaultValue = new Long(prop);
     }
 
     /**
